+++ Src/FluentAssertions/Primitives/NegatedStringStartValidator.cs
@@ -26,8 +26,7 @@ namespace FluentAssertions.Primitives
         {
             get
             {
-                return (stringComparison == StringComparison.CurrentCultureIgnoreCase) ||
-                    (stringComparison == StringComparison.OrdinalIgnoreCase);
+                return stringComparison == StringComparison.OrdinalIgnoreCase;
             }
         }