+++ Src/FluentAssertions/Primitives/StringWildcardMatchingValidator.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Text;
 using System.Text.RegularExpressions;
 
@@ -39,7 +40,11 @@ namespace FluentAssertions.Primitives
 
         private static string ConvertWildcardToRegEx(string wildcardExpression)
         {
-            return "^" + Regex.Escape(wildcardExpression).Replace("\\*", ".*").Replace("\\?", ".") + "$";
+            return "^"
+                + Regex.Escape(wildcardExpression)
+                 .Replace("\\*", ".*", StringComparison.Ordinal)
+                 .Replace("\\?", ".", StringComparison.Ordinal)
+                + "$";
         }
 
         private string CleanNewLines(string input)