+++ Tests/FluentAssertions.Specs/Collections/GenericCollectionAssertionOfStringSpecs.cs
@@ -256,7 +256,7 @@ namespace FluentAssertions.Specs
 
             // Act
             Action action = () => actual.Should().Equal(expected,
-                (a, e) => string.Equals(a, e, StringComparison.CurrentCultureIgnoreCase));
+                (a, e) => string.Equals(a, e, StringComparison.OrdinalIgnoreCase));
 
             // Assert
             action.Should().NotThrow();
@@ -315,7 +315,7 @@ namespace FluentAssertions.Specs
 
             // Act
             Action action = () => actual.Should().Equal(expected,
-                (a, e) => string.Equals(a, e, StringComparison.CurrentCultureIgnoreCase));
+                (a, e) => string.Equals(a, e, StringComparison.OrdinalIgnoreCase));
 
             // Assert
             action