+++ Tests/FluentAssertions.Specs/CultureAwareTesting/CulturedFactAttribute.cs
@@ -1,0 +1,11 @@
+﻿using Xunit;
+using Xunit.Sdk;
+
+namespace FluentAssertions.Specs.CultureAwareTesting
+{
+    [XunitTestCaseDiscoverer("FluentAssertions.Specs.CultureAwareTesting.CulturedFactAttributeDiscoverer", "FluentAssertions.Specs")]
+    public sealed class CulturedFactAttribute : FactAttribute
+    {
+        public CulturedFactAttribute(params string[] cultures) { }
+    }
+}